+++ internal/docs/bloblang.go
@@ -348,6 +348,7 @@ func BloblangMethodsMarkdown() ([]byte, error) {
 		query.MethodCategoryObjectAndArray,
 		query.MethodCategoryParsing,
 		query.MethodCategoryEncoding,
+		query.MethodCategoryGeoIP,
 		query.MethodCategoryDeprecated,
 	} {
 		methods := methodCategory{